+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/constraint/Query.java
@@ -21,6 +21,7 @@ import java.util.function.Function;
 import net.splitcells.dem.data.set.Set;
 import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.data.set.list.Lists;
+import net.splitcells.gel.data.allocation.Allocations;
 import net.splitcells.gel.data.table.attribute.Attribute;
 import net.splitcells.gel.rating.framework.Rating;
 import net.splitcells.gel.rating.rater.Rater;
@@ -38,6 +39,11 @@ import static net.splitcells.dem.data.set.list.Lists.listWithValuesOf;
  * Somehow illegal queries need to be marked via interface.</p>
  */
 public interface Query {
+
+    /**
+     * @return The {@link Allocations} on which the {@link Query} is applied to.
+     */
+    Optional<Allocations> subject();
 
     Query forAll(Attribute<?> args);