+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/SolutionBuilder.java
@@ -35,6 +35,7 @@ import net.splitcells.dem.object.Discoverable;
 import net.splitcells.gel.constraint.Constraint;
 import net.splitcells.gel.constraint.Query;
 import net.splitcells.gel.constraint.QueryI;
+import net.splitcells.gel.data.allocation.Allocations;
 import net.splitcells.gel.data.database.Database;
 import net.splitcells.gel.data.database.Databases;
 import net.splitcells.gel.data.table.attribute.Attribute;
@@ -61,6 +62,7 @@ public class SolutionBuilder implements DefineDemandAttributes, DefineDemands, D
     private Optional<String> name = Optional.empty();
 
     private Constraint constraint;
+    private Allocations allocations;
 
     protected SolutionBuilder(String name) {
         this.name = Optional.of(name);
@@ -80,6 +82,10 @@ public class SolutionBuilder implements DefineDemandAttributes, DefineDemands, D
 
     @Override
     public Problem toProblem() {
+        return problem(allocations, constraint);
+    }
+
+    private void initAllocations() {
         final var problemsDemands = demandsDatabase.orElseGet(() -> {
             // TODO The demands name is a hack.
             final var d = Databases.database(name.orElse(DEMANDS.value()), null, demandAttributes);
@@ -91,22 +97,22 @@ public class SolutionBuilder implements DefineDemandAttributes, DefineDemands, D
             supplies.forEach(supply -> s.addTranslated(supply));
             return s;
         });
-        return problem(
-                allocations(
-                        name.orElse(Solution.class.getSimpleName())
-                        , problemsDemands
-                        , problemsSupplies)
-                , constraint);
+        allocations = allocations(
+                name.orElse(Solution.class.getSimpleName())
+                , problemsDemands
+                , problemsSupplies);
     }
 
     @Override
     public ProblemGenerator withConstraint(Constraint constraint) {
+        initAllocations();
         this.constraint = constraint;
         return this;
     }
 
     @Override
     public ProblemGenerator withConstraint(Function<Query, Query> builder) {
+        initAllocations();
         final var path = Lists.<String>list();
         if (demandsDatabase.isPresent()) {
             path.withAppended(demandsDatabase.get().path());
@@ -114,7 +120,15 @@ public class SolutionBuilder implements DefineDemandAttributes, DefineDemands, D
         if (name.isPresent()) {
             path.withAppended(name.get());
         }
-        return withConstraint(builder.apply(QueryI.query(forAll(Optional.of(discoverable(path))))).currentConstraint());
+        return withConstraint(builder.apply(QueryI.query(forAll(Optional.of(discoverable(path))), allocations)).currentConstraint());
+    }
+
+    @Override
+    public ProblemGenerator withConstraints(List<Function<Query, Query>> builders) {
+        initAllocations();
+        final var root = forAll();
+        builders.forEach(b -> b.apply(QueryI.query(root, allocations)));
+        return withConstraint(root);
     }
 
     @Override