+++ projects/net.splitcells.gel.core/CHANGELOG.md
@@ -51,6 +51,9 @@
   2. Rename `SupplySelector` to `SupplyOfflineSelector`.
   3. Rename `SupplySelectors` to `SupplyOfflineSelectors`.
 ### Minor Changes
+* **2023-04-13 ##170**: Extend Query with optional value containing the Allocations the Query is being applied to.
+  This allows one to create constraints,
+  that depend on the supply and demand Tables of allocations of the given problem.
 * **2023-03-17 ##170**: Introduce `HistoryForDatabase` for debugging purposes.
   An instance can be connected to a given database via the `DatabaseFactory`,
   which can be viewed via the webserver in order to understand how the content of the given database was created.